refactor(Tab): rename to ts
