+++ components/tab/style.ts
