+++ components/tab/types.ts
